refactor: colocate component preset recipes
