+++ internal/agent/PROJECT-UNDERSTANDING.zh-CN.md
@@ -13,7 +13,7 @@
 | Version entity | [`V-PROTO-UI-0005`](../../spec/versions/V-PROTO-UI-0005.yaml) |
 | 工作区实体数 | 394 |
 | Workspace validation issues | 0 |
-| 工作区快照指纹 | `sha256:90a5e9f66cf2918715c171ac948188fe92cdbe655cbd3f10e4e67a7c742e05f0` |
+| 工作区快照指纹 | `sha256:56e7ea946d22150a8aa221cc53d8c786edc9227c3ffcb6ad9a1bdc5a369cc430` |
 | 已发布 release snapshot digest | `sha256:21152b5b6fd27ac2dbffd4bffe71ba0f376c532ccf3d7d074ed3fda846c758d8` |
 
 工作区快照指纹来自按 ID 排序、按当前版本过滤后的实体内容。它用于判断本文是否与当前检出版本一致；它不替代 `V-*` 中记录的不可变发布快照 digest。
