+++ packages/cli/src/registry/shadcn-component-presets.generated.ts
@@ -1,9 +1,10 @@
 /* This file is generated by scripts/components/generate-shadcn-component-presets.mjs. */
-/* Edit packages/prototypes/shadcn/src/component-presets.ts instead. */
+/* Edit the component-local switch/preset.ts or dialog/preset.ts recipe instead. */
 
 export const SHADCN_COMPONENT_PRESET_RECIPES = {
   'shadcn-switch': {
     'kind': 'replaceable-default-part',
+    'placement': 'direct-child',
     'exportName': 'ShadcnSwitch',
     'rootPrototype': 'shadcnSwitchRoot',
     'defaultPartPrototype': 'shadcnSwitchThumb',
@@ -13,6 +14,7 @@ export const SHADCN_COMPONENT_PRESET_RECIPES = {
   },
   'shadcn-dialog': {
     'kind': 'replaceable-default-part',
+    'placement': 'direct-child',
     'exportName': 'ShadcnDialogContent',
     'rootPrototype': 'shadcnDialogContent',
     'defaultPartPrototype': 'shadcnDialogCloseIcon',
