+++ scripts/components/generate-shadcn-component-presets.mjs
@@ -12,7 +12,7 @@ const outputPath = path.join(
 const check = process.argv.includes('--check');
 const rawContent = [
   '/* This file is generated by scripts/components/generate-shadcn-component-presets.mjs. */',
-  '/* Edit packages/prototypes/shadcn/src/component-presets.ts instead. */',
+  '/* Edit the component-local switch/preset.ts or dialog/preset.ts recipe instead. */',
   '',
   `export const SHADCN_COMPONENT_PRESET_RECIPES = ${JSON.stringify(
     shadcnComponentPresets,
